[attmpt] added capping back into intensifier
